bookstore-backend/requirements.txt
jwt~=1.3.1
SQLAlchemy~=2.0.41
fastapi~=0.115.12
phonenumbers~=9.0.6
pydantic~=2.11.5
passlib~=1.7.4
uvicorn~=0.34.3
pydantic-settings~=2.9.1
python-dotenv~=1.1.0
python-json-logger~=3.3.0
psycopg2-binary~=2.9.10
python-jose~=3.5.0
alembic==1.16.1
starlette~=0.46.2
pydantic[email]